Student-GUI-with-SQL-version-2/requirements.txt
pymysql
python-dotenv
uuid-utils
pandas
matplotlib
customtkinter
openpyxl
pyinstaller